[critical performance/security fix] This regexp could stall before for minutes, proportional exponentially to the the size of input (journal entry text.)
The problem was using &.{1,7}; as a pattern for entities was too
general, and the . matched past the ; into the next entity, given a
string of, say, give hundred lt entities right in a row.

If you know how regular expression engines work and how they do
backtracking, you'll see why this sucked major ass.

So now \#?\w+ as the entity content body is used instead of ., which
is actually more accurate.  I found an SGML BNF that also listed
'&u-' Hex4 ';' as a entity type, but I've never seen that.

In any case, all this is unnecessary soon when we support UTF-8.  The
whole reason we had this in the first place was because Russian users
were posting entire posts in entities.  (each letter = 5 characters =
sick)


git-svn-id: http://code.livejournal.org/svn/livejournal@723 1dfba125-df10-0410-8a77-af1a8d6d074f