[unnecessary code]
This was a security thing we did back-when after discovering we had a
few URLs which did an action based on $remote.  The problem is that a
GET request shouldn't do an action, since it can be hidden in an IMG
tag in somebody's post:

[img src="http://lj.com/delete_all_your_shit.bml?confirm=yes" width=1
  height=1]

But I think we've long-since fixed all those cases, so this code is
no longer necessary.


git-svn-id: http://code.livejournal.org/svn/livejournal@762 1dfba125-df10-0410-8a77-af1a8d6d074f