<lj user=avva>: Several fixes in talkpost_do.bml:
1) Use "UTF-8" not "utf-8" as the default mail encoding if the user
provided none (this way, the behaviour is identical in case the user chose
'none' or 'UTF-8' for mail encoding; it's stored as 'UTF-8' in the 'codes'
table since that's the way the conversion library lists it)
2) In case the parent comment is pre-Unicode, translate to UTF-8 not
just the body but also the subject.
3) Besides $textsubject and $cleansubject there needs to be another
variable, $headersubject, which is what gets put in the RFC822 header
(as opposed to the subject also being put inside the body of the email).
The reason is that the whole text (or HTML) of the body is uniformly
converted into $mailencoding before sending, but RFC822 headers aren't
affected by this; so we need to separately convert $headersubject before
calling MIME::Lite->New().


git-svn-id: http://code.livejournal.org/svn/livejournal@981 1dfba125-df10-0410-8a77-af1a8d6d074f