<lj user=avva>: This patch does two things:
1. Correctly encodes Subject: fields in emails as per RFC2047.
Note it uses MIME::Words , which should be on all our servers (I think
it's part of the standard distribution, but it's better to check).
2. Inserts a meta tag with encoding information in the beginning of
the HTML part of the email (some mail readers aren't smart enough to
apply the Content-type: information already there in the MIME header
to HTML attachments, so this gives them a clue).

Tested.

<lj user=bradfitz>:
the lj-bml-init/INSTALL changes


git-svn-id: http://code.livejournal.org/svn/livejournal@994 1dfba125-df10-0410-8a77-af1a8d6d074f