[yeah, i'm slow]Date: Sun, <b>03 Mar 2002</b> 16:05:29 +0000 From: Rob Martin <gooner_@hotmail.com> To: opiummmm@livejournal.com, bradfitz@livejournal.com Subject: community/shared issue
... [snip] ...

(I'll take this opportunity to remind you of the previous community bug(?) I
wrote to you about before where a user can't access their post in a
community when they have no longer have posting access, so they can't
edit/delete it.  I imagine this is the exact same thing as the recent thing
with the /suggestions/ community)

Cheers!

-rob


git-svn-id: http://code.livejournal.org/svn/livejournal@1044 1dfba125-df10-0410-8a77-af1a8d6d074f