<lj user=avva>: This patch fixes a bug in editjournal_do.bml which prevents people from modifying the current mood from the one in the drop-down list to a custom mood (w/o changing anything else in the entry).
The bug was that prop_current_moodid is deleted from the req hash if
a custom field prop_current_mood is present. It needs to be emptied,
not deleted, so that the protocol code will see it and delete the
existing moodid of the entry (if it was there).

Tested.


git-svn-id: http://code.livejournal.org/svn/livejournal@1358 1dfba125-df10-0410-8a77-af1a8d6d074f