<lj user=avva>: Here's the same patch with anti-autovivifying checks added. I also fixed the same problem elsewhere in ljviews.pl in this patch (specifically where it deals with opt_nclinks).
The original description of the patch:

This patch addresses a few minor security holes in friend group names,
improves the way filtering works in links, and makes public friend
groups actually work.

<lj user=signe> submitted a patch which prevented non-owner from
seeing filter values in skiplinks (thus verifying that the group name
if valid) and made /username/friends/group/ syntax work in addition
to /username/friends/group .

<lj user=avva> reviewed and tested that, then added support for
public groups, and rewrote skiplinks code to use the group by name
if it was specified by name, thus making skiplinks less cryptic.
Also presence of default view is no longer detectable by skiplinks
after this patch.

What it does:
ljlib.pl - LJ::get_friend_items(): passed value of $filter is used
even if the
remote is not the owner (needed for public groups; security check
shifts higher to the view creator).

ljviews.pl - create_view_friends():
1. Only consider using $FORM{'filter'} if remote==owner;
2. If a group is specified by name, only convert it to a $filter
value if it's either public or remote==owner;
3. When building skiplinks, propagate the group name in the link if
it was there. Only propagate the filter value if it was requested in the
URL (so that if we built $filter from the default view group, it
won't be reflected in the skiplinks).

ljprotocol.pl - don't allow creating groups with last character '/' .


Tested.
Here's the same patch with anti-autovivifying checks added. I also fixed the
same problem elsewhere in ljviews.pl in this patch (specifically where
it deals with opt_nclinks).

The original description of the patch:

This patch addresses a few minor security holes in friend group names,
improves the way filtering works in links, and makes public friend
groups actually work.

<lj user=signe> submitted a patch which prevented non-owner from
seeing filter values in skiplinks (thus verifying that the group name
if valid) and made /username/friends/group/ syntax work in addition
to /username/friends/group .

<lj user=avva> reviewed and tested that, then added support for
public groups, and rewrote skiplinks code to use the group by name
if it was specified by name, thus making skiplinks less cryptic.
Also presence of default view is no longer detectable by skiplinks
after this patch.

What it does:
ljlib.pl - LJ::get_friend_items(): passed value of $filter is used
even if the
remote is not the owner (needed for public groups; security check
shifts higher to the view creator).

ljviews.pl - create_view_friends():
1. Only consider using $FORM{'filter'} if remote==owner;
2. If a group is specified by name, only convert it to a $filter
value if it's either public or remote==owner;
3. When building skiplinks, propagate the group name in the link if
it was there. Only propagate the filter value if it was requested in the
URL (so that if we built $filter from the default view group, it
won't be reflected in the skiplinks).

ljprotocol.pl - don't allow creating groups with last character '/' .


Tested.


git-svn-id: http://code.livejournal.org/svn/livejournal@1370 1dfba125-df10-0410-8a77-af1a8d6d074f