<lj user=mart>: syndication userinfo prettiness
I plan to do the following:
* Change the title to say 'Syndication Account' rather than 'Syndicated
Account', since it seems to me to read better that way. (Which is, of
course, a translation text change, and'll be provided as an importable texty
thing)
* Remove Name and Website and replace with "Syndicated From", which has the
account name as a link to the site, with the orange XML button thingy after
it.
* Hide friends/friendof, since it doesn't matter who watches a syndicated
journal. Also, of course, remove the calls to getfriends.
* Hide Account Type, Clients used, Journal entries and Comments (from
mode=full), because they're all irrelevant to syndicated journals. (and, in
the case of 'Journal Entries', eventually old stuff'll be deleted so this
number won't be useful)
* Change the paragraph at the top to reference a different translation text
which, by default, will read something along the lines of "This account is
updated with articles from the below mentioned site. You can add this
account to your friends list to make these articles appear on your friends
page."


git-svn-id: http://code.livejournal.org/svn/livejournal@1405 1dfba125-df10-0410-8a77-af1a8d6d074f