documentation update from <lj user=jproulx>
Automatic capability document generation, along with instructions on how
to define caps/classes in ljconfig.pl

Added the ability to include examples for the XML-RPC protocol calls, so
that developers can get a quick glimpse of what they should be expecting.
I will do the same for the flat protocol documentation on the next sweep.

- Opi


git-svn-id: http://code.livejournal.org/svn/livejournal@1427 1dfba125-df10-0410-8a77-af1a8d6d074f