<lj user=bradfitz>: I asked Avva to add support for a new db role type "cluster[n]lite" to act like "cluster[n]slave", except with old log2 and talk2... very heavily used tables, and very small.  By putting them on their own machines, we could get some good cache wins on some very cheap machines.
<lj user=avva>:
Here's an updated patch which implements the lite role. The changes to
ljlib.pl are to get_recent_items() as before, plus implementation of
'onlysubjects' in get_talktext2().

The changes to talkread.bml are: eliminate loading subjects in the
large query which loads all comments, and consider using the lite db
for this query if possible. Then find only those comments that
actually are displayed with just their subjects, when threading, and
load subjects only for those comments, instead of loading subjects for
everything as before. This should speed up talkread.bml on large
comment pages.

I tried to test as exhaustively as possible, with lots of large
threads, mixed display of subject-only and full comments, etc. Seems
to work fine.


git-svn-id: http://code.livejournal.org/svn/livejournal@1433 1dfba125-df10-0410-8a77-af1a8d6d074f