Add support to update-db.pl to group multiple small layers (such as themes) together in one file.
The syntax to mark a file as a multi-file is a + sign after the
layer type in s2layers.dat.

theme+ means a file with a bunch of themes.

Then, in the file, new layers start with:

#NEWLAYER: some/redist/value

This makes CVS mangling easier.  We'll eventually have tons of themes,
and they don't all need their own file.


git-svn-id: http://code.livejournal.org/svn/livejournal@1728 1dfba125-df10-0410-8a77-af1a8d6d074f