Trying to fix: http://zilla.livejournal.org/776
I'm able to reproduce the blank / bug with a dynamic mod_perl, with perl_module loaded last,
and this patch fixes that.  (set_handlers shouldn't be called this late, I believe?  so it
only works in some configs.  but isn't supported like it was being used?)

However, I can't reproduce the talkread doubling problem which was reported previously with
this patch.  I've tried a whole bunch of apache configs... dynamic, static, at end, at
beginning, with and without other modules which would affect things...


git-svn-id: http://code.livejournal.org/svn/livejournal@1985 1dfba125-df10-0410-8a77-af1a8d6d074f