ignore $u->{'lang'}, which for a million users is "EN" but never used anywhere.
ljprotocol ignored it before, since it didn't select that field, but
then we changed to using LJ::load_user(), so suddenly everybody
was language "EN", which isn't a valid language code since it's capitalized.

so, with this patch we totally ignore $u->{'lang'} and always use browselang.

this fixes the "Server has a message for you:  ?lang?"


git-svn-id: http://code.livejournal.org/svn/livejournal@2407 1dfba125-df10-0410-8a77-af1a8d6d074f