community memories should be editable by community maintainer http://zilla.livejournal.org/show_bug.cgi?id=217
While I was at it, I switched FORM to GET/POST and fixed up a translation
string.

The proper behavior here was confusing.  Whitaker and I spent a good hour plus
discussing the relationship between authas and can_manage:  if evan can authas
martine and martine manages the community lj_dev, should evan be able to
manage lj_dev?  It would be nice, but it doesn't work in the current design of
things, where authas and management are blurred together.


git-svn-id: http://code.livejournal.org/svn/livejournal@2707 1dfba125-df10-0410-8a77-af1a8d6d074f