Generator lacks font settings http://zilla.livejournal.org/show_bug.cgi?id=1113
Mart writes:
This patch adds font setting support to Generator, assuming the font-family
suggestion it had before worked. (I just replaced it with a chunk of code from
one of the other layouts)

Evan adds:
just a newline, so the CSS looks better.
(and I removed all the whitespace changes).


git-svn-id: http://code.livejournal.org/svn/livejournal@2762 1dfba125-df10-0410-8a77-af1a8d6d074f