This should unbreak lots of stuff when visiting old-style talkread URLs (pre-clusterisation). Parent/thread/linkbar links should now work, as well as polls embedded in those entries.
Explanation: LJ::Talk::init put in $item->{'ditemid'} a copy of
the URL's itemid. We initialise $ditemis in talkread.bml from
$item->{'ditemid'}. Then later we fix up $init->{'ditemid'} in case
the URL was old-style (we put the new-style public itemid there),
but we neglect to fix the $ditemid variable in the same way, and it's
used to initialise $talkurl used for all relative URLs, and for
other stuff such as polls which were also converted to new-style URLs.

Thanks to Whitaker for a painstaking review of this one-liner and
its possible implications.


git-svn-id: http://code.livejournal.org/svn/livejournal@3165 1dfba125-df10-0410-8a77-af1a8d6d074f