Support syndicated Atom feeds and improve support of RSS syndication.
1) New library: cgi-bin/parsefeed.pl is now responsible for parsing feeds.
It uses XML::RSS for RSS feeds and XML::Parser with custom handlers for
Atom feeds (version 0.3 currently). It returns the parsed feed in a
uniform format to the caller.

2) RSS parsng has been enhanced to support publication times in <pubDate>
and unique ids of items in <guid>.

3) New logprop, syn_id, is used for storing unique ids of syndicated items
(always present as <id> in Atom, and sometimes present as <guid> in RSS).

4) synsuck.pl has been modified to use the new library, parsefeed.pl, and
to support setting eventtime from the feed rather than always set it
to the time of sucking.


git-svn-id: http://code.livejournal.org/svn/livejournal@3399 1dfba125-df10-0410-8a77-af1a8d6d074f