First example of the brand new and supercool plural parser, we really need a better name for this stuff.
Closing this Zilla Bug:
Memories page: Bad translation string
http://zilla.livejournal.org/show_bug.cgi?id=213

Check out Brad's commit to read more about the technical implementation.
http://www.livejournal.com/community/changelog/1117563.html

My blurb explaining how to use it.
The translation system now supports the ability to add multiple plural forms of the word
given different rules in a languge.  This functionality is much like the plural support
in the S2 styles code.  To use this code you must use the BML::ml function and pass
the number of items as one of the variables.  To make sure that you are allowing the
utmost compatibility for each language you should not hardcode the placement of the
number of items in relation to the noun.  Let the translation string do this for you.
A translation string is in the format of, with num being the variable storing the number of items.
=[[num]] [[?num|singular|plural1|plural2|pluralx]]


git-svn-id: http://code.livejournal.org/svn/livejournal@3439 1dfba125-df10-0410-8a77-af1a8d6d074f