we're going to start depending on clustertrack2 for lots more stuff and kill off userusage.timecheck in the process.
userusage.timecheck generates lots of global writes which we hardly
ever use.  instead, we need to convert the two remaining users of
userusage.timecheck over to use clustertrack2.timeactive:

-- bin/maint/stats.pl  (for active user stats)
-- bin/ljumover.pl  (for determining active users to move)

Stats will require backing off and rescheduling if/when it can't get a
cluster reader handle, otherwise stats for that day would be way off.

ljumover.pl will involve making sure every user on a cluster is in the
the clustertrack2 table.  well, no!  I thought it would, but instead
what we can do is: if the ljumover.pl move command includes a +active
declaration, then we do a clustered read on clustertrack2 to find
users to move.  if there's no active rule, then we just do a query on
user as it does now.  then the only additional change to ljumover.pl
is making sure it does a LJ::load_userids_multiple on the user table
to get the info it didn't get from the clustered clustertrack2
(columns: statusvis and user)


git-svn-id: http://code.livejournal.org/svn/livejournal@3480 1dfba125-df10-0410-8a77-af1a8d6d074f