HUGE (~3x) speedup in html cleaner, even after all the other speedups.  The long-word-breaker regexp was fucking horrid.
the speed-up is very similar to the one I did previous today, which
was: "Assume the post/comment has no HTML and we can clean it
trivially."  That yielded its own huge speedup.

This speed-up is: instead of assuming the comment/post has a long word
in it and finding the perfect place to break it while not messing up
UTF-8 multi-byte characters, instead just look for some 40 bytes of
non-whitespace (which is fast) and only then do a closer inspection
with the slow regexp.

For instance, consider a typical comment:

"Yeah, I agree.  Good plan."

The first regular expression isn't going to find 40 bytes of
contiguous non-whitespace.

But when it finds something like:

"YEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEEAH,
that's totally awesome."

Then it'll get only get run on the "YEEE..." part, and not all the other
fine text.


git-svn-id: http://code.livejournal.org/svn/livejournal@3505 1dfba125-df10-0410-8a77-af1a8d6d074f