http://zilla.livejournal.org/1659
The S2 function get_url is used to construct a URL to a journal.  If you
pass it a UserLite S2 object, it returns a URL that points to the right
place.  If you give it just a username, however, it points to the wrong
place if the username is a community.

This patch fixes that by making it check to see what kind of account it
is and appropriately modifying the URL for communities.


git-svn-id: http://code.livejournal.org/svn/livejournal@3697 1dfba125-df10-0410-8a77-af1a8d6d074f