minor optimization.  (again, the $LJ::DYNAMIC_LJUSER thing like from interests.bml the other day)
but this isn't as big a win as that turned out to be, since the LJ::load_user()
that gets called from LJ::ljuser() will generally be cached in this case.

just wanted to do it because it was easy, and talkread.bml is where we
spend most our CPU time, so every bit helps a lot in here.


git-svn-id: http://code.livejournal.org/svn/livejournal@3746 1dfba125-df10-0410-8a77-af1a8d6d074f