New console command: allow_open_proxy.
This takes an IP address.  If that IP address is found in the openproxy table
and the user sending the command has the "allowopenproxy" privilege and the
IP is an open proxy, then it is marked as "clear" for the next 24 hours with
a note that this was done manually by someone.


git-svn-id: http://code.livejournal.org/svn/livejournal@3823 1dfba125-df10-0410-8a77-af1a8d6d074f