[int: todo list security] http://zilla.livejournal.org/1109
This really isn't much of a security issue, but eh.  Anyway, there's a problem
right now such that if you have the friends group priv-todo-* (where * is
add, edit, or delete), it doesn't matter if anybody is IN that group, the mere
presence of the group means that everybody you friend has access to do that
action with your todo list.

Problem was that the SQL statement to figure out permissions was missing the
appropriate part of the WHERE clause.  Added that in.


git-svn-id: http://code.livejournal.org/svn/livejournal@3845 1dfba125-df10-0410-8a77-af1a8d6d074f