http://zilla.livejournal.org/1597
New members of FriendsPage to determine what kind of view it is (friends or
friendsfriends and eventually maybe friendof).  Also gives the ability to
determine if a private filter is in use and employs this to put a smart
title on the view by default.  Patch by <lj user="mart">.


git-svn-id: http://code.livejournal.org/svn/livejournal@3916 1dfba125-df10-0410-8a77-af1a8d6d074f