First version of the Journal Backup (jbackup) tool.  This tool is a Perl script that enables people to maintain local copies of their journals, including all entries and all comments ever made to said journals.
This also works on communities that you are maintainer of.  So you can now
create backup copies of those, as well.

The tool creates a GDBM file database and also features two modes of getting
data out of the database: XML and HTML.  (And raw, but you probably won't need
that...)

See the built in help (command line option --help) for help on how to use it.


git-svn-id: http://code.livejournal.org/svn/livejournal@3946 1dfba125-df10-0410-8a77-af1a8d6d074f