New function: LJ::do_to_cluster.
It is a meta-function that takes as a parameter a single subref.  It then
picks a random starting point within @LJ::CLUSTERS and begins iterating
through up to 15 of them, calling the subref with the cluster id as a
parameter.

Designed to be used for things such as buffering emails, where you want a
random cluster to be used for buffering up the job.  This function stops
executing as soon as the subref returns 1 or it fails on 15 clusters.

This patch also modifies the three spots that use cmd_buffer_add on a
random cluster to use this new function.


git-svn-id: http://code.livejournal.org/svn/livejournal@4021 1dfba125-df10-0410-8a77-af1a8d6d074f