New console command: expunge_userpic &lt;picid&gt; Required privilege: siteadmin:userpics, siteadmin:*
This console command will take a user picture and mark it as expunged
(state=X) which will cause the system to refuse to serve this picture.
In addition, the picture is no longer counted towards the user's quota.
In effect, it no longer exists.

This also calls a hook, "expunge_userpic", if said hook exists.  This
can be used by sites that have off-site cacheing and need to notify them
that the picture should be removed from the cache.

Expunges are also logged in statushistory, under the keyword "expunge_userpic".

Also, no longer serve userpics for suspended users.  They're still there,
and they will be available again if the user is unsuspended, but they are
unavailable for the duration of the suspension.


git-svn-id: http://code.livejournal.org/svn/livejournal@4027 1dfba125-df10-0410-8a77-af1a8d6d074f