New features: thread deletion, thread freezing.
The former puts a new checkbox on the delcomment.bml page so that you can
choose to delete a comment thread.  (A comment and all the replies and
replies of replies, etc.)

Thread freezing provides the abilities for journal administrators to freeze
a thread, thereby disallowing further replies to it.  Threads can be frozen
and unfrozen at will.  The S2 "Comment" class also now has a "frozen" member
indicating whether the comment is frozen or not.  Icons by <lj user="jproulx">.


git-svn-id: http://code.livejournal.org/svn/livejournal@4084 1dfba125-df10-0410-8a77-af1a8d6d074f