I b0rked the site.  Whoops.
The dversion 5 codepath in this code was using an inefficient query that
was causing the global to start doing huge scans through all of the data
looking for information.

Uh, that's bad.  So it's fixed.  Had to add a join back in, which is what
I removed that caused it to break.  Now that I recombined the queries so
they join again, they work.  (Moving to dversion 6 is good, now!)


git-svn-id: http://code.livejournal.org/svn/livejournal@4197 1dfba125-df10-0410-8a77-af1a8d6d074f