New console command: change_journal_status [username] [newstatus]
Statuses are 'normal', 'memorial', and 'locked'.  They affect the accounts
in these ways:

Memorial accounts cannot have new entries posted in them.

Locked accounts cannot:
 - post/receive comments
 - post entries
 - login (big one here...they can't log in)
 - edit friends, edit groups, edit entries

However, locked accounts can still access data -- view entries, comments,
and export such with the protocol, etc.

Note: This will potentially be in flux in the future, but this is the
behavior as of this commit.


git-svn-id: http://code.livejournal.org/svn/livejournal@4337 1dfba125-df10-0410-8a77-af1a8d6d074f