EXPERIMENTAL // FOR EXAMPLE USE ONLY:
Demonstrate how to alter the security on entries using the jbackup format.
It's really very simple given the infrastructure provided by jbackup.

Should help client authors to get an idea of how to properly write clients
that will download a user's entire journal and then alter the security and
carefully reupload it.


git-svn-id: http://code.livejournal.org/svn/livejournal@4355 1dfba125-df10-0410-8a77-af1a8d6d074f