add LJ::User->selectrow_array proxy method, because:
LJ::alloc_user_counter was buggy when $u->set_dbcm was used, setting
the $u to an uncached, raw $dbcm.  In that case, the SELECT LAST_INSERTID()
part of alloc_user_counter would use a separate handle from the one that
$u->do above it did, and the return value of alloc_user_counter would
always be 0.

Hasn't been a problem so far in normal code, since $u->set_dbcm is new
and only used for bizarre cases, but it affected a mover of ours.


git-svn-id: http://code.livejournal.org/svn/livejournal@4501 1dfba125-df10-0410-8a77-af1a8d6d074f