Small change that doesn't really do justice to the massive shift in how the user's counter information works.
Basically, counter information is now stored on the global.  Counter info
that was clustered is now in the 'usercounter' table.

The old 'counter' table on clusters is now deprecated and should not be
used for any purpose.

Reason for the change: previously, with the master/master setup, if the
master is switched and another counter is allocated while the first side is
still catching up, this can lead to a conflict (same number allocated twice)
and causing replication to break.

The solution used by command buffer, having the a/b sides do even/odd
increases in the numbers so as not to bump into eachother, was ruled out
because counters shouldn't go up by 2.

This is tested and small, and doesn't cause any known issues.  I looked
around in the code for other spots that use the counter table directly,
but didn't find any... here's hoping.


git-svn-id: http://code.livejournal.org/svn/livejournal@4557 1dfba125-df10-0410-8a77-af1a8d6d074f