Very buried bug:
    * if a user posts in a community, specifying a user picture, and
    * someone viewing this post on a friends page has use_shared_pic off, and
    * the community and poster are on different clusters, and
    * the picture is not currently in memcache (fallen out, or no memcache)

If all of the above is true, the user viewing their friends page (S2 only)
would see no user picture by the post by the person in the community.  It
would appear that they specified to not user a user picture at all, as there
wasn't even an IMG tag or a wrong link.

Fixed by keeping track of what user we should use to load the picture,
because it changes depending on a few things.


git-svn-id: http://code.livejournal.org/svn/livejournal@4656 1dfba125-df10-0410-8a77-af1a8d6d074f