Synopsis: it was reported that the userpic deactivation code was hitting up the wrong userpics.  This is true; the query for finding a post from 2 months ago was wrong.
Fixed with this patch, change the + to a -.

Logic: (hopefully this makes sense to someone other than me)

Let's have ourselves a little graph of the normal situation:

NOW            TIME=0
| a  t       b      |
|....t..............|

So time is progressing to the left... we want to find something to the
right of point t.  Anything in the b area, which is a query looking for
things to the right of time t:

    WHERE logtime < (UNIX_TIMESTAMP() - 86400 * 60)

We want "to the right of t", and right is in the past, so, "where logtime
is less than t".  We define t as 60 days ago.  Makes sense.

But in this case, we have to use rlogtime.  So that changes the graph some:

TIME=0            EOT
|    a    t    b    |
|.........t.........|

Note the swapped TIME=0 position, and EOT (End of Time, max 32 bit int) on
the right.  Time is still progressing to the left, and the two chunks are
still a and b.  HOWEVER, a is now a much bigger segment than b... but that
doesn't matter.  We still want to find things that are contained in b.

However, since t is going towards 0, and b is to the right of it, we want
to look for things where rlogtime is *greater* than the rlogtime of time t.

    WHERE rlogtime > ($LJ::EndOfTime - UNIX_TIMESTAMP() - 86400 * 60)

So basically, the problem was that we were incorrectly converting the
current time to time 't'.  We want two months in the past, not two months
in the future, which is what we were getting.

Alright, so that's clear as mud.


git-svn-id: http://code.livejournal.org/svn/livejournal@4675 1dfba125-df10-0410-8a77-af1a8d6d074f