If a user used to be paid and set a domain alias up at that time, then let their account expire, they have no way of removing the domain alias.
This isn't a problem except in the case where the same person creates
another paid account later that they want to use the same domain.

Had a support request for this, here's solution I went with:

-- modify.bml gets a "remove" button if the user has a journaldomain set,
   but no longer has the userdomain cap.

-- paid users never see the remove button, because <lj user="jproulx"> and
   I both agree it's not the best interface in the world.

This also removes an unnecessary DELETE on domains in the case that a user
used to have a domain alias and  clears it.


git-svn-id: http://code.livejournal.org/svn/livejournal@4788 1dfba125-df10-0410-8a77-af1a8d6d074f