There's been a bug that causes behavior where you select a layout, click change, then on the response page Generator is selected.  Quite annoying.
I was able to reproduce on my home machine and found that it was because
I had styles whose layout layerid was 0.  However, I was unable to find
out what exactly causes this, though I tried for quite a while to reproduce
a case that would set the layout layerid to 0.  :-/

If I set up an artificial test case by manually setting a wizard-uniq
named style's layout layerid to 0, then this patch fixes it.  The patch
makes sense to me anyway, and I'm not sure why we didn't do it before.

Changes:

-- When changing layout layerid, we search for a style named
   "wizard-layoutname" and use that.  But once we select it we don't verify
   that its layout is actually "layoutname".  Set it explicitly.

-- Remove a duplicate call to LJ::S2::set_style_layers()


git-svn-id: http://code.livejournal.org/svn/livejournal@4791 1dfba125-df10-0410-8a77-af1a8d6d074f