Cluster s2compiled2.
Basically, the new table is now keyed on (userid, layerid) and contains
data that is compressed when it's put in.  (Tests show this reduces data
size by about 75%.)

Still need a script to import the data to the new tables... that should
be coming along tomorrow.

Tested this a bunch, it works for me.

Patch is a joint effort by Brad and myself.  I'm pretty lost in S2 land.  :-)


git-svn-id: http://code.livejournal.org/svn/livejournal@4847 1dfba125-df10-0410-8a77-af1a8d6d074f