Rework the loading layers logic with regards to memcache:
* If a layer is process cached (within the last 60 seconds) it is
considered good and we use it as is.  (This is fine since the bug with not
unregistering user layers is fixed...and even then, it'd still be only 60
seconds with old data before it flushes.)

* After that, we fall back to memcache.  If a layer is in memcache, we
consider it to be valid.  We also now delete from memcache every time we
update the s2compiled2 table, keeping everything in sync.

* Finally, it falls back to the database, but it makes sure to load the
data into memcache.

So, in a nutshell, system styles end up process cached for 60 seconds, and
user layers end up cached in memcache, and theoretically we'll never have
to really hit the database.


git-svn-id: http://code.livejournal.org/svn/livejournal@4858 1dfba125-df10-0410-8a77-af1a8d6d074f