Add userlog table that is useful for general purpose status logging.  Right now, it only stores when an entry is deleted.
The table stores the initiator of the event so that we are able to look back
and say "this person deleted the entry".  This works for communities and
individual journals and the actions are logged on the account that lost the
entry.

The basic format is [ action, actiontarget ]. An action is just a textual
description of what happened (up to 30 characters) and the target is
optional but is likely some numeric value.

There is no way to view this information on the site, yet, as it's more useful
as admin information.


git-svn-id: http://code.livejournal.org/svn/livejournal@4922 1dfba125-df10-0410-8a77-af1a8d6d074f