Don't show that annoying alert box telling users that they had typed in a username but they hadn't selected the username radio box when their browser was the one that "typed" a username.
So this is the annoying case:

-- you're logged in

-- the talk form properly selects the radio for "Logged in user: bob"

-- your browser fills in the other LJ username box for "sam" (another person who had used your computer)

-- you write your comment, hit post, then you get dumb alert

Now, instead, we track whether or not you ever focused onto that other
username text box and only show the alert if so.

Also, clear that other username before sending to server when that
radio's not checked, so server doesn't hit you with the same error.


git-svn-id: http://code.livejournal.org/svn/livejournal@4928 1dfba125-df10-0410-8a77-af1a8d6d074f