Noticed this with disk quota of 5_000 MiB:
If disk quota is set to a value requiring over 12 bits, then left shifts
of 20 bits will overflow and result in a number way-too-small.

However, if you make it an integer multiplication of 1 << 20, then Perl
will automatically convert the number to a big integer and do the right
thing.

Changing for shifts of 10 as well, to be safe.


git-svn-id: http://code.livejournal.org/svn/livejournal@4941 1dfba125-df10-0410-8a77-af1a8d6d074f