[int: security issue]
Fix a bug that allows people to get the email address on their account set
to anything and have it be validated.

Simply verify that the email address on the authaction being used is the
same as the address on the account.  Otherwise, have the user request a
new one.


git-svn-id: http://code.livejournal.org/svn/livejournal@5005 1dfba125-df10-0410-8a77-af1a8d6d074f