David and I want to kill as much configuration as possible.
This whole %LJ::INNODB_DB is stupid and fragile.  It can be
auto-detected in the future.

For now, consolidate the few callers (er, checkers) of %LJ::INNODB_DB
all into the $u->is_innodb function, which in the future will
auto-discover if it's InnoDB or not and cache it (per clusterid) for
the duration of the process and/or in memcached.


git-svn-id: http://code.livejournal.org/svn/livejournal@5914 1dfba125-df10-0410-8a77-af1a8d6d074f