Remove the join on user from the IM account search query in multisearch.bml.
This isn't as scary as it perhaps first appears; the most common case is that only one userid will be
returned, and worst case is that 25 userids will be returned and all get loaded at once, probably right from
memcache, using load_userids.

Motivation for this is that eventually the 'user' table will be moving to its own
DB role, meaning that it will no longer be possible to join on it.


git-svn-id: http://code.livejournal.org/svn/livejournal@6296 1dfba125-df10-0410-8a77-af1a8d6d074f