Add links list to UserLite in similar vein to the links list on EntryLite. This includes the same stuff that appears on the link toolbar on userinfo.bml, including a link to post in the journal if the user has permission to do that.
This, like its companion in EntryLite, has hardcoded English link captions.
Ideally this should really be using the translation system, but it's a bit
trickier since it should really be using the style language rather than
the remote user's active language.

Someone will have to decide where to put this stuff in the various system layouts.


git-svn-id: http://code.livejournal.org/svn/livejournal@6299 1dfba125-df10-0410-8a77-af1a8d6d074f