magic userpic keyword for mood mapping http://rt.livejournal.org/Ticket/Display.html?id=195
that is, now (or soon) the userpic for an entry will be determined
by the post's mood, if there's userpic keyword defined on a post.

that is, you can put the keyword "sad" on a userpic, and if you ever
use the mood "sad" on a post, it uses your sad userpic.


git-svn-id: http://code.livejournal.org/svn/livejournal@6344 1dfba125-df10-0410-8a77-af1a8d6d074f