Add a flag to EntryLite which indicates when the "text" member ABSOLUTELY MUST be printed trusted. Most layers should always be using print_text(), but some users want to do crazy parsing/whatever of the "text" member and need a way to know when that must be bypassed and a trusted print performed.
The way this flag is set is a little lame, but is based on the kind of stupid logic
people have been putting in their S2 layers. At least now it's united in one place
and can be changed later to something more clever.

If you're currently doing something resembling this in your S2 layer, please switch
to using this new flag instead of your own wacky logic. It'll be better for you
in the long run.


git-svn-id: http://code.livejournal.org/svn/livejournal@6631 1dfba125-df10-0410-8a77-af1a8d6d074f