[int: ads] http://rt.livejournal.org//Ticket/Display.html?id=1267
Fix link text.  Is there a better way of saying "a LiveJournal" than hard-coding
"LiveJournal" into the code?  SITENAMESHORT doesn't seem appropriate, since
that's supposed to be the site's name, not the name of a journal.  Is there a
variable we can use for the term for the site's journals?


git-svn-id: http://code.livejournal.org/svn/livejournal@6703 1dfba125-df10-0410-8a77-af1a8d6d074f