removing _clear_u_req_cache, as it'd break singleton rules.  there are no callers of it currently, but I don't want there to be any in the future.
instead, you should upgrade an object or mark a singleton as dirty so
on next use it gets lazily upgraded in place.  but i never want two $u
variables in a process with the same userid but different contents.

we can already upgrade.  if we need to mark a $u as dirty, we can add
that in the future, but it hasn't been necessary.



git-svn-id: http://code.livejournal.org/svn/livejournal@7068 1dfba125-df10-0410-8a77-af1a8d6d074f