Get shot of the stupid link text properties in A Sturdy Gesture.
These were originally added before the system had such text in core, but they're a pain
in the butt because the layer needs to be updated every time a new link is added.

This patch changes the layout to use the system-supplied caption. The old text properties
are still consulted for backwards-compatibility, but they're now considered to be
purely customization-oriented (for user layers) rather than as the definitive source of
text for these buttons. Consequently, the layout defines none of them by default.



git-svn-id: http://code.livejournal.org/svn/livejournal@7098 1dfba125-df10-0410-8a77-af1a8d6d074f