Content resize/realign (CRR) for S2 Clean and Simple.
This is pretty ugly looking, but that's because I didn't want to
break the functionality of the "border width" property that already
exists.  So I made the border width property only affect the top
and bottom borders unless the content width is 100%.

I tested this a bunch but I may have missed something.  For those
of you who have local installs, feel free to test this and let me
know if I messed something up.



git-svn-id: http://code.livejournal.org/svn/livejournal@7176 1dfba125-df10-0410-8a77-af1a8d6d074f