http://rt.livejournal.org/Ticket/Display.html?id=1902
This is the start of the "make S2 wizard options consistent between styles" project.  Basically,
we're just changing property and property group names and order.  This way, a given property will
always be under the same property group in all styles, and people will hopefully be able to find
things more easily.



git-svn-id: http://code.livejournal.org/svn/livejournal@7246 1dfba125-df10-0410-8a77-af1a8d6d074f