viewer_is_friend() and viewer_is_member() functions for S2.
Both of them do the same thing internally, but make the distinction between communities
(which use "friends" in a weaker sense) and other kinds of journals (where the "friends"
relationship generally implies trust, or whatever.)

Was going to add a third function that doesn't care about journal type, but couldn't
think of a good name for it. Back on Zilla I proposed having this separated out to
viewer_can_read, viewer_is_trusted, etc in preparation for a view/watch/trust separation,
but I don't think that's planned anymore?



git-svn-id: http://code.livejournal.org/svn/livejournal@7310 1dfba125-df10-0410-8a77-af1a8d6d074f