http://rt.livejournal.org/Ticket/Display.html?id=1798
S2 Disjointed!  Thanks to <lj user="kunzite1"> for making it.

I tested a lot of stuff and added in all of the options that should be necessary to bring it up to par
with S1 Disjointed.  But it's quite possible I missed things.  Feel free to test this if you want to.

Also, I updated the dat file to add this new style, and I fixed a typo in Core.



git-svn-id: http://code.livejournal.org/svn/livejournal@7540 1dfba125-df10-0410-8a77-af1a8d6d074f