Change around the T-Mobile regex to reject certain image files. Previously, we only allowed them through if they were named a certain way, but that's pretty unreliable (and changed pretty much right after we committed the last revision of this). Now, let's strip out the images we know are "bad".
Of course, they could just change the file names again, but I think it's better to let through a "bad" image and *all* the user's files than to strip out the ads *and* some of the user's files. If that's the case, we can just update the regex again later on.

http://rt.livejournal.org/Ticket/Display.html?id=90


git-svn-id: http://code.livejournal.org/svn/livejournal@7642 1dfba125-df10-0410-8a77-af1a8d6d074f