Next layout-wide addition: adding custom external stylesheet and custom CSS box support.  Some layouts already had this, some only have one, some have none.  So now all layouts will have both, and the descriptions/notes will all be consistent (except for Bloggish which is a special case).  This is the start.
For the record, I'm taking the $*custom_css property out of print_stylesheet() and putting it in its own print
in Page::print().  This is so the order of stylesheets can be: layout CSS, custom external stylesheet, custom CSS
from the box in the wizard.  I think that makes the most sense, it terms of CSS precedence.



git-svn-id: http://code.livejournal.org/svn/livejournal@7766 1dfba125-df10-0410-8a77-af1a8d6d074f