Adding custom CSS properties (custom external stylesheet, option to include layout's stylesheet, and custom CSS box) to Core, along with a Page function to do the logic and print the appropriate stylesheets: Page::print_stylesheets().
Please let me know if this function should be global instead; I couldn't decide since print_stylesheet() is global, but
I think this new one should be in Page (makes more sense to me).

Thank you to <lj user="murklinstest"> for pointing out the need for a flag to turn off a layout's stylesheet when including
a custom external stylesheet.  And thank you to <lj user="mart"> for pointing out that this should all be in Core.  :)

Layout modifications will come next.



git-svn-id: http://code.livejournal.org/svn/livejournal@7882 1dfba125-df10-0410-8a77-af1a8d6d074f