Very old bug that people may or may not have noticed:
When the external stylesheet is generated for an S2 style, it appends a cache-buster in the form of a modtime.
The modtime is that of either the layer in the style with the greatest modtime or the modtime of the style itself,
whichever is greater.

However, it appears that the code was actually getting the modtime of the user's current style in the second case,
as opposed to the modtime of the style being viewed.  As a result, if someone was using ?s2id to view a style and
then changed the layers in that style, they were getting the original layout/theme's stylesheet instead of the new
one's (until a hard refresh).  I believe that this would only have been noticable to users who modify the layers
in custom styles often and who are using layouts with external stylesheets, which is why it seems it's never been
noticed until now.

In conclusion, have this check the modtime of the style being viewed, not the user's current style.


git-svn-id: http://code.livejournal.org/svn/livejournal/branches/customize@11800 1dfba125-df10-0410-8a77-af1a8d6d074f