Bug ID: 61865
Code that auto-populates the vertical table with data from %LJ::VERTICAL_TREE.  It will add any verticals
that are defined but not in the table, and will remove any verticals that are in the db but not defined.

Also removed a bit of code from the test so that it works given that we have initial data in the table now.


git-svn-id: http://code.livejournal.org/svn/livejournal/branches/vertical-content@13034 1dfba125-df10-0410-8a77-af1a8d6d074f