Bug ID: 62925
Don't explicitly disconnect log DB handles at the end of
a run of LJ::EventLogSink::Database.  This code is an
old remnant of when the DB logging happened in Apache
land, but now it breaks LJ::EventLogRecord/Sink in the
case that multiple sinks subscribe to an Event and
Database runs first.



git-svn-id: http://code.livejournal.org/svn/livejournal/branches/vertical-content@13220 1dfba125-df10-0410-8a77-af1a8d6d074f